requirements.txt
flask
flask-cors
flask-socketio
requests
pycryptodome
